tests/wpt/web-platform-tests/trusted-types/block-string-assignment-to-frame-src.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./support/helper.sub.js"></script>

<meta http-equiv="Content-Security-Policy" content="require-trusted-types">
<body>
<script>
  //URL assignments don't throw
  test(t => {
    var url = TrustedURL.create(URLS.safe);

    var d = document.createElement('frame');
    d.src = url;
    assert_equals("" + d.src, URLS.safe);
  }, "src = TrustedURL.create().");

  test(t => {
    var url = TrustedURL.unsafelyCreate(URLS.safe);

    var d = document.createElement('frame');
    d.src = url;
    assert_equals("" + d.src, URLS.safe);
  }, "src = TrustedURL.unsafelyCreate().");

  //String assignment throws
  test(t => {
    var d = document.createElement('frame');
    assert_throws(new TypeError(), _ => {
      d.src = "Fail.";
    });
  }, "'src = string' throws.");
</script>
